Fixing issue 220
https://github.com/vitch/jScrollPane/issues/220

When calculating whether the element is above/below viewport
settings.verticalGutter is used instead of settings.horizontalGutter.
The same problem is with with calculating whether the element is to the
left/right viewport.

So when I set { verticalGutter: 10 } and use scrollToElement function
the pane is scrolled 10px beneath needed though verticalGutter should
not effect it at all