OverlayTrack added
In some cases, we don't want panel to cut-off to show trackbar. If trackbar is positioned absolute then it can also overflow on top of panel without changing it's width attributes. Otherwise causes list items with background color set, not to fill container but leave a gap on side of the trackbar.