Add symbolic links to styles and script folders
The example site at jscrollpane.kelvinluck.com stopped working after
4d5ee25 because I just serve up a copy of the repository and moving
the files changed the location of all of the things that were being
served.

This change should mean the  example site can be reinstated and
hopefully will continue to work at the old URLs (I'll point the
webserver serving it to the `examples` directory rather than the root).